create ui folder and add generic textarea and input components
