+++ src/components/ui/textarea/textarea.scss
@@ -1,0 +1,29 @@
+@import '../../../styles/extends';
+
+.textarea {
+  @extend %input;
+
+  resize: none;
+
+  &--large {
+    @extend %input--large;
+  }
+
+  &--small {
+    @extend %input--small;
+  }
+}
+
+.textarea-character-count {
+  align-items: baseline;
+  color: var(--color-text-faded);
+  display: flex;
+  font-size: 16px;
+  justify-content: flex-end;
+  margin-bottom: 6px;
+  width: 100%;
+}
+
+.textarea-number-characters {
+  color: var(--color-text);
+}
